Making pause button and Game Over functional
