+++ FrontEnd/Script/Gameplay/scriptGameplay.js
@@ -5,6 +5,8 @@ import SpaceDreadnought from "../Models/Bosses/scriptSpaceDreadnought.js";
 import AssetLoader from "./scriptAssetLoader.js";
 import * as PATHS from "./scriptConstants.js";
 import InputManager from "./scriptInputManager.js";
+import gameOver from "./scriptGameOver.js";
+import { values } from "./scriptDOM.js";
  
 const canvas = document.getElementById("gameCanvas");
 const contex = canvas.getContext("2d");
@@ -33,122 +35,131 @@ let autoFireState = {
     autoFireInterval: 120 // ms between shots when holding
 };
 
+export const gameState = {
+    isPaused: false,
+    isGameOver: false,
+};
+
 const gameLoop = () => {
-    contex.clearRect(0, 0, canvas.width, canvas.height);
 
-    input.update();
+    if (!gameState.isPaused && !gameState.isGameOver) {
+        contex.clearRect(0, 0, canvas.width, canvas.height);
 
-    if (input.keys.left && player.position.x > 0) player.moveLeft();
-    if (input.keys.right && player.position.x < canvas.width - player.width) player.moveRight();
-    if (input.keys.up && player.position.y > 0) player.moveUp();
-    if (input.keys.down && player.position.y < canvas.height - player.height) player.moveDown();
+        input.update();
 
-    // --- Rotação (por botões Q/E ou LB/RB) ---
-    if (input.keys.rotateLeft) player.rotateLeft();
-    if (input.keys.rotateRight) player.rotateRight();
+        if (input.keys.left && player.position.x > 0) player.moveLeft();
+        if (input.keys.right && player.position.x < canvas.width - player.width) player.moveRight();
+        if (input.keys.up && player.position.y > 0) player.moveUp();
+        if (input.keys.down && player.position.y < canvas.height - player.height) player.moveDown();
 
-    // --- Tiro (Espaço ou ZR/R2/RT): single on press, continuous when held ---
-    const now = Date.now();
-    if (input.keys.spacePressed) {
-        // immediate single shot on press
-        player.shoot(bullets);
-        autoFireState.holding = true;
-        autoFireState.holdStart = now;
-        autoFireState.lastShot = now;
-    }
+        // --- Rotação (por botões Q/E ou LB/RB) ---
+        if (input.keys.rotateLeft) player.rotateLeft();
+        if (input.keys.rotateRight) player.rotateRight();
 
-    if (input.keys.space) {
-        // if still holding and hold delay passed, fire at interval
-        if (autoFireState.holding) {
-            if (now - autoFireState.holdStart >= autoFireState.holdDelay) {
-                if (now - autoFireState.lastShot >= autoFireState.autoFireInterval) {
-                    player.shoot(bullets);
-                    autoFireState.lastShot = now;
-                }
-            }
-        } else {
-            // if space becomes held without edge (rare), initialize holding state
+        // --- Tiro (Espaço ou ZR/R2/RT): single on press, continuous when held ---
+        const now = Date.now();
+        if (input.keys.spacePressed) {
+            // immediate single shot on press
+            player.shoot(bullets);
             autoFireState.holding = true;
             autoFireState.holdStart = now;
             autoFireState.lastShot = now;
         }
-    } else {
-        // released
-        autoFireState.holding = false;
-    }
 
-    bullets.forEach(b => {
-        b.update();
-        b.draw(contex);
-    });
-
-    enemies.forEach(e => {
-        e.update(player, bullets, canvas);
-        e.draw(contex);
-    });
-
-    spaceDreadnought.draw(contex);
-    spaceDreadnought.update(player, bullets);
-
-    collisionManager.entities = [player, spaceDreadnought,  ...enemies, ...bullets];
-    collisionManager.update();
-
-    // Integrate repulsion velocities for player and enemies (simple impulse integration + damping)
-    const integrateEntities = [player, ...enemies];
-    integrateEntities.forEach(ent => {
-        if (ent.vx || ent.vy) {
-            // apply small movement from impulse
-            ent.position.x += ent.vx;
-            ent.position.y += ent.vy;
-
-            // damping - reduce velocity gradually (lower damping => bouncier)
-            ent.vx *= 0.75;
-            ent.vy *= 0.75;
-
-            // clamp tiny velocities to zero
-            if (Math.abs(ent.vx) < 0.01) ent.vx = 0;
-            if (Math.abs(ent.vy) < 0.01) ent.vy = 0;
-            
-            // Prevent entities (especially the player) from leaving the canvas due to impulses
-            const minX = 0;
-            const minY = 0;
-            const maxX = canvas.width - ent.width;
-            const maxY = canvas.height - ent.height;
-
-            // clamp X
-            if (ent.position.x < minX) {
-                ent.position.x = minX;
-                if (ent.vx < 0) ent.vx = 0; // stop outward velocity
-            } else if (ent.position.x > maxX) {
-                ent.position.x = maxX;
-                if (ent.vx > 0) ent.vx = 0;
+        if (input.keys.space) {
+            // if still holding and hold delay passed, fire at interval
+            if (autoFireState.holding) {
+                if (now - autoFireState.holdStart >= autoFireState.holdDelay) {
+                    if (now - autoFireState.lastShot >= autoFireState.autoFireInterval) {
+                        player.shoot(bullets);
+                        autoFireState.lastShot = now;
+                    }
+                }
+            } else {
+                // if space becomes held without edge (rare), initialize holding state
+                autoFireState.holding = true;
+                autoFireState.holdStart = now;
+                autoFireState.lastShot = now;
             }
+        } else {
+            // released
+            autoFireState.holding = false;
+        }
 
-            // clamp Y
-            if (ent.position.y < minY) {
-                ent.position.y = minY;
-                if (ent.vy < 0) ent.vy = 0;
-            } else if (ent.position.y > maxY) {
-                ent.position.y = maxY;
-                if (ent.vy > 0) ent.vy = 0;
+        bullets.forEach(b => {
+            b.update();
+            b.draw(contex);
+        });
+
+        enemies.forEach(e => {
+            e.update(player, bullets, canvas);
+            e.draw(contex);
+        });
+
+        spaceDreadnought.draw(contex);
+        spaceDreadnought.update(player, bullets);
+
+        collisionManager.entities = [player, spaceDreadnought,  ...enemies, ...bullets];
+        collisionManager.update();
+
+        // Integrate repulsion velocities for player and enemies (simple impulse integration + damping)
+        const integrateEntities = [player, ...enemies];
+        integrateEntities.forEach(ent => {
+            if (ent.vx || ent.vy) {
+                // apply small movement from impulse
+                ent.position.x += ent.vx;
+                ent.position.y += ent.vy;
+
+                // damping - reduce velocity gradually (lower damping => bouncier)
+                ent.vx *= 0.75;
+                ent.vy *= 0.75;
+
+                // clamp tiny velocities to zero
+                if (Math.abs(ent.vx) < 0.01) ent.vx = 0;
+                if (Math.abs(ent.vy) < 0.01) ent.vy = 0;
+                
+                // Prevent entities (especially the player) from leaving the canvas due to impulses
+                const minX = 0;
+                const minY = 0;
+                const maxX = canvas.width - ent.width;
+                const maxY = canvas.height - ent.height;
+
+                // clamp X
+                if (ent.position.x < minX) {
+                    ent.position.x = minX;
+                    if (ent.vx < 0) ent.vx = 0; // stop outward velocity
+                } else if (ent.position.x > maxX) {
+                    ent.position.x = maxX;
+                    if (ent.vx > 0) ent.vx = 0;
+                }
+
+                // clamp Y
+                if (ent.position.y < minY) {
+                    ent.position.y = minY;
+                    if (ent.vy < 0) ent.vy = 0;
+                } else if (ent.position.y > maxY) {
+                    ent.position.y = maxY;
+                    if (ent.vy > 0) ent.vy = 0;
+                }
+            }
+        });
+
+        for (let i = bullets.length - 1; i >= 0; i--) {
+            const b = bullets[i];
+            if (!b.active ||
+                b.position.x < 0 || b.position.x > canvas.width ||
+                b.position.y < 0 || b.position.y > canvas.height) {
+                bullets.splice(i, 1);
             }
         }
-    });
 
-    for (let i = bullets.length - 1; i >= 0; i--) {
-        const b = bullets[i];
-        if (!b.active ||
-            b.position.x < 0 || b.position.x > canvas.width ||
-            b.position.y < 0 || b.position.y > canvas.height) {
-            bullets.splice(i, 1);
+        for (let i = enemies.length - 1; i >= 0; i--) {
+            if (!enemies[i].active) enemies.splice(i, 1);
         }
-    }
 
-    for (let i = enemies.length - 1; i >= 0; i--) {
-        if (!enemies[i].active) enemies.splice(i, 1);
+        player.draw(contex);
     }
-
-    player.draw(contex);
+    
     requestAnimationFrame(gameLoop);
 }
 
@@ -173,6 +184,17 @@ AssetLoader.preload(Array.from(new Set(assets)), (progress) => {
     // console.log(`Loading assets: ${(progress * 100).toFixed(0)}%`);
 }).then(() => {
     console.log("✅ Todos os assets carregados!");
+    // Listen for player death events dispatched by game objects
+    window.addEventListener('playerGameOver', () => {
+        gameState.isGameOver = true;
+        // read user from localStorage and call gameOver UI
+        try {
+            const user = JSON.parse(localStorage.getItem('user'));
+            gameOver(user, values);
+        } catch (e) {
+            console.error('Failed to open game over UI', e);
+        }
+    }, { once: true });
     gameLoop();
 }).catch(err => {
     console.error("Falha ao pré-carregar assets:", err);
