+++ FrontEnd/Script/Gameplay/scriptPause.js
@@ -1,11 +1,18 @@
+import { gameState } from "./scriptGameplay.js";
+
 const pauseButton = document.getElementById("pauseButton");
 
 const user = JSON.parse(localStorage.getItem("user"));
 if (!user) window.location.href = "../../enter.html";
 
-pauseButton.addEventListener("click", () => {
+if (!pauseButton) {
+    console.warn('pauseButton element not found');
+} else {
+    pauseButton.addEventListener("click", () => {
+
+    gameState.isPaused = true;
     if (document.getElementById("pauseScreen")) return;
-    
+
     console.log("Pause button clicked ✅");
     const pauseScreen = document.createElement("div");
     pauseScreen.id = "pauseScreen";
@@ -79,7 +86,10 @@ pauseButton.addEventListener("click", () => {
     // EVENT LISTENERS
     resumeButton.addEventListener("click", () => {
         setDisabledAll(true, resumeButton);
-        setTimeout(() => pauseScreen.remove(), 150);
+        setTimeout(() => {
+            pauseScreen.remove();
+            gameState.isPaused = false;
+        }, 150);
     });
 
     saveSoundsButton.addEventListener("click", async () => {
@@ -120,8 +130,9 @@ pauseButton.addEventListener("click", () => {
         });
     });
 
-    // Append elementos
-    pauseContent.append(pauseTitle, resumeButton, soundSlider, musicSlider, saveSoundsButton, exitButton);
-    pauseScreen.appendChild(pauseContent);
-    document.body.appendChild(pauseScreen);
-});
+        // Append elementos
+        pauseContent.append(pauseTitle, resumeButton, soundSlider, musicSlider, saveSoundsButton, exitButton);
+        pauseScreen.appendChild(pauseContent);
+        document.body.appendChild(pauseScreen);
+    });
+}
